Backends

Here are the functions related to the backends. Users usually would just need to use the set_GPUbackend() or set_CPUbackend().

HighSeas.get_backend — Function

  get_backend()

Return the current backend used to perform the calculations.

Example

get_backend()

>>> "CPUBackend"

HighSeas.set_GPUbackend — Function

  set_GPUbackend()

Set the backend to the available GPU. This is decided by the installed packages, i.e. if CUDA.jl is installed (and imported) then CUDA will be used. Only import one JuliaGPU package per script. If no JuliaGPU package is loaded, this method will return a LoadError.

See the available GPU platforms by running get_available_GPUplatforms().

Example

using HighSeas
using CUDA

set_GPUbackend()

>>> CUDA with CUDA.DeviceMemory will now be used

using HighSeas
using AMDGPU

set_GPUbackend()

>>> AMDGPU with AMDGPU.Runtime.Mem.HIPBuffer will now be used

Notes

Some GPU backends support UnifiedMemory (e.g. CUDA, Metal). The default will always be DeviceMemory but users can choose UnifiedMemory by running set_GPUbackend("unified")

HighSeas.set_CPUbackend — Function

  set_CPUbackend()

Set the backend to CPU

Example

set_CPUbackend()

>>> CPU will now be used

HighSeas.get_available_platforms — Function

  get_available_platforms()

Get the list of supported platforms that can be used for running the simulations

Example

get_available_platforms()

>>> ["CPU", "CUDA", ...]

HighSeas.get_available_GPUplatforms — Function

  get_available_GPUplatforms()

Get the list of supported GPU platforms that can be used for running the simulations

Example

get_available_GPUplatforms()

>>> ["CUDA", ...]