Detectors

Here are all the detectors that can be used to detect if an event is happening or not. When a detector is used, the simulation state is saved in the experiments' output folder.

Simple detectors

These are simple detectors that either do nothing (necessary for code homogeneity) or very simple things (i.e. detect the event but do nothing with the data)

HighSeas.EmptyDetector — Type

EmptyDetector <: AbstractDetector

Detector that does nothing

HighSeas.SimpleDetector — Type

SimpleDetector{AbstractState, AbstractStepper} <: AbstractDetector

Simple detector that just prints a message and logs the event in the simulation.log file

Fields

eventStart::Bool – Flag used to define if the event has started or not

eventN::Int – Event number

minVThresh::Float64 – Minimum slip rate threshold to define the END of the event

maxVThresh::Float64 – Maximum slip rate threshold to define the START of the event

state::AbstractState – State of the simulation

stepper::AbstractStepper – Stepper used in the simulation

log_file::String – Path of where to log possible console outputs

Notes

When used, the detector saves the AbstractSavers of the simulation at the end of each event.

The dector has an associated functor that can be used to detect the event

Examples

...

detector = SimpleDetector(minVThresh, maxVThresh, experiment, algorithm) # define the detector

...

detector(savers) # Detect a possible event given the thresholds

Complex detectors

These are more advanced detectors that detect events and do something (i.e. define a catalog entry)

HighSeas.CatalogDetector — Type

CatalogDetector{AbstractState, AbstractStepper, AbstractCatalog, AbstractArray{Float64}, AbstractArray{Int8}} <: AbstractDetector

Detector used to detect an event and calculate the associated values and fill a catalog

Fields

eventStart::Bool – Flag used to define if the event has started or not

eventN::Int – Event number

minVThresh::Float64 – Minimum slip rate threshold to define the END of the event

maxVThresh::Float64 – Maximum slip rate threshold to define the START of the event

state::AbstractState – State of the simulation

stepper::AbstractStepper – Stepper used during simulation

material::AbstractMaterial – Material of the simulation

catalog::AbstractCatalog – Catalog used to saved the calculated values

dx_start::AbstractArray{Float64} – Values of slip at the start of the event. These are pre-allocated temporary matrices used to calculate the different quantities in the catalog

tau_start::AbstractArray{Float64} – Values of τ at the start of the event. These are pre-allocated temporary matrices used to calculate the different quantities in the catalog

slip::AbstractArray{Float64} – Values of slip at the end of the event. These are pre-allocated temporary matrices used to calculate the different quantities in the catalog

stressdrop::AbstractArray{Float64} – Stress drop at the start of the event. These are pre-allocated temporary matrices used to calculate the different quantities in the catalog

x::AbstractArray{Float64} – x coords of the grid. These are pre-allocated temporary matrices used to calculate the different quantities in the catalog

y::AbstractArray{Float64} – y coords of the grid. These are pre-allocated temporary matrices used to calculate the different quantities in the catalog

ruptured_nodes::B – Mask of which nodes are being ruptured

time_start::Float64 – Start of the rupture

last_event_time::Float64 – last rupture time (to calculate the intertime event)

cell_area::Float64 – Area of the cells

log_file::String – Path of where to log possible console outputs

Notes

When used, the detector saves the AbstractSavers of the simulation at the end of each event.

The dector has an associated functor that can be used to detect the event

When using GPU, it is possible to define where the temporary matrices reside using gpu_id

Examples

...

detector = CatalogDetector(minVThresh, maxVThresh, experiment, algorithm; gpu_id::Int=0) # define the detector

...

detector(savers) # Detect a possible event given the thresholds