Experiments

Here are the available experiments

HighSeas.BP4QDExp — Type

BP4QDExp{AbstractArray{Float64}, AbstractMaterial, AbstractDomain, AbstractState, AbstractCatalog}

Create the experiment object for the BP4QD benchmark. N_events sets the size of the Catalog.

Fields

material::AbstractMaterial – Material of the experiment

domain::AbstractDomain – Domain of the experiment

start_time::String – Start time of the experiment

outpath::String – Where to save the output of the simulations

Vpl::Float64 – Velocity of the plate (m/s)

Vr::Float64 – Reference velocity (m/s)

Vi::Float64 – Initial velocity (m/s)

Vnu::Float64 – Nucleation patch velocity (m/s)

lengthscales::NamedTuple – Calculated lengthscales (Lb, Linf, L/Linf)

a::AbstractArray{Float64} – Matrix of a values for rate and state

b::AbstractArray{Float64} – Matrix of b values for rate and state

tau0::AbstractArray{Float64} – Matrix of initial tau values for rate and state (Pa)

si0::AbstractArray{Float64} – Matrix of initial si for rate and state (Pa)

state::AbstractState – State of the simulation (this is not θ of the rate and state!!)

catalog::AbstractCatalog – Catalog of the simulation

Notes

When defining the experiment from scratch (i.e. not making a copy) the simulation state (::AbstractState) and catalog (::AbstractCatalog) are automatically created

When using GPUs, it is possible to set the gpu_id that defines where the temporary matrices reside

If an experiment must be resumed by a backup, it is possible to pass a LoadedStep object

When the experiment is created, the following file structure is generated

.
└── Timestamp/
    └── Backend/
        └── simulation.log

Where the simulation.log will have the input sheet + eventual appended messages by other components (i.e. the AbstractDetectors).

Examples

BP4QDExp(input_dict::Dict, material::AbstractMaterial, domain::AbstractDomain, n_events::Int; gpu_id=0) – Create a fresh experiment

BP4QDExp(input_dict, material, domain, n_events, output_dir, loadedstep::LoadedStep; gpu_id=0) – Resume an experiment from a given step