Grids, Geometries and Domains

Here are all the grids and geometries that can be used to compose the domain

Grids

HighSeas.Grid — Type

Grid <: AbstractGrid

Create a grid object.

Fields

cell_sizex::Float64 – Size of the cell along x (m)

cell_sizey::Float64 – Size of the cell along y (m)

cell_area::Float64 – Area of the cell

W::Float64 – width of the entire domain

L::Float64 – length of the entire domain

n_elementsx::Int – number of elements on the x dimension

n_elementsy::Int – number of elements on the y dimension

X::StepRangeLen – X range

Y::StepRangeLen – Y range

x::AbstractArray{Float64} – x coordinates array

y::AbstractArray{Float64} – y coordinates array

Examples

Grid(input_dict::Dict) – Create a grid from an input dict

HighSeas.PowerGrid — Type

PowerGrid <: AbstractPowerGrid

Create a grid object in which cells are square and the number of elements in the grid is a power of 2.

Fields

cell_sizex::Float64 – Size of the cell along x (m)

cell_sizey::Float64 – Size of the cell along y (m)

cell_area::Float64 – Area of the cell (m2)

W::Float64 – width of the entire domain

L::Float64 – length of the entire domain

domain_powerx::Int – Power of 2 along the x dimension

domain_powery::Int – Power of 2 along the y dimension

n_elementsx::Int – number of elements on the x dimension

n_elementsy::Int – number of elements on the y dimension

X::StepRangeLen – X range

Y::StepRangeLen – Y range

x::AbstractArray{Float64} – x coordinates array

y::AbstractArray{Float64} – y coordinates array

Notes

To define the closest power of 2 for x and y, a rounding operation is performed as follows: round(log2(L/cellsize)), round(log2(W/cellsize))

Examples

– PowerGrid(input_dict::Dict) – Create a dictionary of input values

Geometries

HighSeas.RectangleFault — Type

RectangleFault <: AbstractFault

Create a rectangular fault

Fields

Wf::Float64 – half width of the fault

Lf::Float64 – half length of the fault

dLO::AbstractArray{Int8} – Mask indicating the loading area

dCR::AbstractArray{Int8} – Mask indicating the creeping region

Notes

When using GPUs, it is possible to decide where the masks reside using gpu_id

Examples

RectangleFault(input_dict::Dict, grid::AbstractGrid; gpu_id=0) – Create a rectangle fault from an AbstractGrid

HighSeas.RectanglePatch — Type

RectanglePatch(input_dict, grid) <: AbstractPatch

Create a rectangular slip patch

Attributes

w::Float64 – half width of the rate weakening zone

l::Float64 – half length of the rate weakening zone

h::Float64 – Buffer between RW and RS

dRW::AbstractArray{Int8} – Mask indicating the rate weakening zone

dRS::AbstractArray{Int8} – Mask indicating the rate strengthening zone

dTR::AbstractArray{Int8} – Mask indicating the transition between RS and RW

Notes

h can be set to 0, i.e. entails that there is no transition zone between RW and RS.

When using GPUs, it is possible to decide where the masks reside using gpu_id

Examples

RectanglePatch(input_dict::Dict, grid::AbstractGrid; gpu_id=0) – Create a rectangle fault from an AbstractGrid

HighSeas.CustomPatch — Type

CustomPatch <: AbstractPatch

Create a custom slip patch object using coordinates of a polygon.

Fields

w::Float64 – half width of the rate weakening zone

l::Float64 – half length of the rate weakening zone

h::Float64 – Buffer between RW and RS

dRW::AbstractArray{Int8} – Mask indicating the rate weakening zone

dRS::AbstractArray{Int8} – Mask indicating the rate strengthening zone

dTR::AbstractArray{Int8} – Mask indicating the transition between RS and RW

Notes

If no width or length are defined, then the bounding box of the polygon will be used

If no buffer is defined, then no buffer will be used.

If a buffer size, then it will be defined as the area between the polygon and a scaled up version of it such that the perpendicular distance between the same point in the two shapes is = h

If a buffer polygon is defined then it will be used to create the buffer and h will be calculated as the difference between the bounding box half-length of the buffer and custom patch.

When using GPUs, it is possible to decide where the masks reside using gpu_id

Examples

CustomPatch(points::AbstractMatrix{Float64}, grid::AbstractGrid) – Create patch from polygon, no buffer

CustomPatch(points::AbstractMatrix{Float64}, grid::AbstractGrid; w=w, l=l, h=h) – Create patch from polygon, specifying w, l and buffer

CustomPatch(points::AbstractMatrix{Float64}, grid::AbstractGrid, buffer_points::AbstractMatrix{Float64}) – Create patch from polygon and buffer polygon

HighSeas.EmptyNucleation — Type

EmptyNucleation <: AbstractNucleation

Create an empty nucleation object i.e. no nucleation patches in the domain.

Fields

xi::Float64 – x center of the nucleation zone

yi::Float64 – y center of the nucleation zone

wi::Float64 – width of the nucleation zone

li::Float64 – length of the nucleation zone

dNU::AbstractArray{Int8} – Mask indicating inside the nucleation zone

dFD::AbstractArray{Int8} – Mask indicating everything outside the nucleation zone

HighSeas.RectangleNucleation — Type

RectangleNucleation <: AbstractNucleation

Create a rectangular nucleation object i.e. a rectangular (or square) nucleation patch in the domain.

Fields

xi::Float64 – x center of the nucleation zone

yi::Float64 – y center of the nucleation zone

wi::Float64 – width of the nucleation zone

li::Float64 – length of the nucleation zone

dNU::AbstractArray{Int8} – Mask indicating inside the nucleation zone

dFD::AbstractArray{Int8} – Mask indicating everything outside the nucleation zone

Notes

When using GPUs, it is possible to decide where the masks reside using gpu_id

Examples

RectangleNucleation(input_dict::Dict, grid::AbstractGrid; gpu_id::Int=0) – Create a nucleation patch using an input dict

Domains

HighSeas.Domain — Type

Domain <: AbstractDomain

Create the domain object.

Attributes

grid::AbstractGrid – Grid object

fault::AbstractFault – Fault object

patch::AbstractPatch – Rate weakening patch object

nucleation::AbstractNucleation – Nucleation patch object

Notes

– If no nucleation object is given then it is assumed that there is no nucleation patch.

Examples

Domain(grid, fault, patch) – Create a domain without a nucleation patch

Domain(grid, fault, patch, nucleation) – Create a domain with a nucleation patch