Saving and reading

Here are objects and functions related to saving and reading simulation results.

Saving

Saving data is managed automatically when solving. The user can choose what to save using the appropriate AbstractSavers. The data will be saved as in JLD2 format for all savers except for the SnapshotSaver which will be saved as png or svg depending on the backend (GPU will save as .png, CPU .svg)

HighSeas.StepSaver — Type

StepSaver <: AbstractSaver

Save the state (dx, V, theta and tau) of the simulation at the given step.

Fields

outpath::String – Where to save the output

every::Int – Save every nth step

data::AbstractState – State of the simulation to save

stepper::AbstractStepper – Stepper used to regulate the simulation

Examples

StepSaver(every::Int, experiment::AbstractExperiment, algorithm::AbstractAlgorithm)

Notes

When using AbstractDetectors, every is ignored and the step is saved at the end of each detection

When defining the saver, outpath is set from the experiment

HighSeas.CatalogSaver — Type

CatalogSaver <: AbstractSaver

Save the catalog of the simulation at the given step.

Fields

outpath::String – Where to save the output

every::Int – Save every nth step

data::AbstractCatalog – Catalog of the simulation to save

stepper::AbstractStepper – Stepper used to regulate the simulation

Examples

CatalogSaver(every::Int, experiment::AbstractExperiment, algorithm::AbstractAlgorithm)

Notes

When using AbstractDetectors, every is ignored and the step is saved at the end of each detection

When defining the saver, outpath is set from the experiment

HighSeas.SamplerSaver — Type

SamplerSaver <: AbstractSaver

Save the samplers used during the simulation at the given step.

Fields

outpath::String – Where to save the output

every::Int – Save every nth step

data::Vector{<:AbstractSampler} – Samplers to save

stepper::AbstractStepper – Stepper used to regulate the simulation

Examples

SamplerSaver(every::Int, experiment::AbstractExperiment, algorithm::AbstractAlgorithm)

Notes

When using AbstractDetectors, every is ignored and the step is saved at the end of each detection

When defining the saver, outpath is set from the experiment

HighSeas.SnaptshotSaver — Type

SnaptshotSaver <: AbstractSaver

Save AbstractPlotters during the simulation at the given step.

Fields

outpath::String – Where to save the output

every::Int – Save every nth step

data::AbstractPlotter – State of the simulation to save

stepper::AbstractStepper – Stepper used to regulate the simulation

Examples

SnaptshotSaver(every::Int, experiment::AbstractExperiment, algorithm::AbstractAlgorithm)

Notes

When using AbstractDetectors, every is ignored and the step is saved at the end of each detection

When defining the saver, outpath is set from the experiment

Reading

HighSeas.loadData — Function

loadData(input::String)
loadData(input::String, n_events::Int)

Load AbstractSaver data

Input

input::String – Input path

n_events::Int – Number of events to load

Notes

n_events is used to load catalogs up to a certain row

HighSeas.loadSSH — Function

loadSSH(url::String, username::String, private_file::String, public_file::String, path::String)

Load AbstracSaver data using SSH

Input

url:String – ssh url

username:String – username

private_file:String – ssh private key

public_file:String – ssh pub key

path::String – Path of the data to read

HighSeas.readSheet — Function

  readSheet(path_file::String)

Read the input sheet

Inputs

path_file::String – Path to the input sheet file to read

Outputs

Dictionary of values

Notes

To use this function the input file must contain the following

+ fract             : increase for higher precisions (higher=slower)
+ tollo             : lower threashold for the error calculations
+ tollup            : upper threashold for the error calculations
+ cs                : shear wave velocity [m/s]
+ rho               : density [kg/m³]
+ nu                : poisson ratio
+ a                 : rate and state constant
+ b                 : rate and state constant
+ aRS               : "a" value for Rate Strengtening (RS)
+ si0               : value of constant (or initial) normal stress
+ Dc                : Characteristic length [m]
+ Vpl               : Velocity of the plate [m/s]
+ Vi                : Initial velocity [m/s]
+ Vr                : Reference velocity [m/s]
+ Vnu               : Nucleation velocity [m/s]
+ fr                : Reference friction
+ cellsize          : size of the cell (cell is always square) 
+ W                 : Width of the simulation domain
+ L                 : Length of the simulation domain
+ Wf                : Half-width of the fault
+ Lf                : Half-length of the fault
+ w                 : Half-width of the Rate Weakening (RW) patch
+ l                 : Half-length of the Rate Weakening (RW) patch
+ h                 : Buffer size between RW and RS (can be zero)
+ xi                : x center of the nucleation patch
+ yi                : y center of the nucleation patch
+ wi                : Half-width of the nucleation patch (can be zero)
+ li                : Half-length of the nucleation patch (can be zero)
+ tf                : final time to reach

HighSeas.LoadedStep — Type

LoadedStep <: AbstractLoadedObject

Loaded step

Fields

state::AbstractState – Saved state of the step

step::Int – Step

time::Float64 – Time of the step

HighSeas.LoadedSamplers — Type

LoadedSamplers <: AbstractLoadedObject

Loaded samplers

Fields

n_samplers::Int – Number of samples in the SamplerSaver

info::Dict – Information of the saved samplers

samplers::Vector – Samplers