Laws

Here are all the available AbstractLaw objects that can be used to solve rate and state

HighSeas.GoverningEquations — Type

GoverningEquations <: AbstractGoverningEquation

Commodity object used to group the governing equations of the simulation

Fields

ratelaw::AbstractRateLaw – Rate law

stresslaw::AbstractStressLaw – Stress law

statelaw::AbstractStateLaw – State law

Rate laws

HighSeas.ExplicitRate — Type

ExplicitRate <: AbstractRateLaw

Create an ExplicitRate object to calculate explicitly the rate from the Rate and State equation.

Fields

eta::Float64 – Radiation damping coefficient

fr::Float64 – Friction reference coefficient

Vr::Float64 – Reference slip velocity

Dc::Float64 – Characteristic length

V_mem::AbstractArray{Float64} – This is the memory matrix in which the result is saved

Notes

When using GPUs, it is possible to decide where V_mem resides using gpu_id

Examples

ExplicitRate(experiment::AbstractExperiment; gpu_id::Int=0) – explicit rate law for the current experiment

HighSeas.LinearizedRate — Type

LinearizedRate <: AbstractRateLaw

Create an LinearizedRate object to calculate the rate using a first order Taylor expansion.

Fields

eta::Float64 – Radiation damping coefficient

fr::Float64 – Friction reference coefficient

Vr::Float64 – Reference slip velocity

Dc::Float64 – Characteristic length

V_mem::AbstractArray{Float64} – This is the memory matrix in which the result is saved

Notes

When using GPUs, it is possible to decide where V_mem reside using gpu_id

Examples

LinearizedRate(experiment::AbstractExperiment; gpu_id::Int=0) – linear rate law for the current experiment

HighSeas.HybridRate — Type

HybridRate <: AbstractHybridRateLaw

Create an HybridRate object to calculate the rate using either the explicit or linearized method.

Fields

threshold::Float64 – Velocity threshold to decide which law to use

explicit::AbstractRateLaw – First law to use in case the V is <= than the thresh

linear::AbstractRateLaw – Second law to use in case the V is > than the thresh

State laws

HighSeas.AgeingLaw — Type

AgeingLaw <: AbstractStateLaw

Create an AgeingLaw object to use the ageing law for calculating the state in the rate and state simulation

Fields

Dc::Float64 – Dc value

theta_mem::AbstractArray{Float64} – This is the memory matrix in which the result is saved

Notes

When using GPUs, it is possible to decide where theta_mem resides using gpu_id

Examples

AgeingLaw(experiment::AbstractExperiment; gpu_id::Int=0) – AgeingLaw rate law for the current experiment

HighSeas.SlipLaw — Type

SlipLaw(experiment::AbstractExperiment)

Create an SlipLaw object to use the slip law for calculating the state in the rate and state simulation

Fields

Dc::Float64 – Dc value

theta_mem::AbstractArray{Float64} – This is the memory matrix in which the result is saved

Notes

When using GPUs, it is possible to decide where theta_mem resides using gpu_id

Examples

SlipLaw(experiment::AbstractExperiment; gpu_id::Int=0) – SlipLaw law for the current experiment

Stress laws

HighSeas.StressFFT — Type

StressFFT <: AbstractStressLaw

Analytical stress-slip relationship in the Fourier/wavenumber domain.

Fields

akdhat::AbstractArray{Float64} – akdhat values

taur::AbstractArray{Float64} – Matrix used to store the calculated stress values from fft

tau::AbstractArray{Float64} – Matrix used to store the final stress (tau = tau0+taur)

F::AbstractArray{ComplexF64} – Matrix used to store the matrix mul result in the wavenumber domain

dx_hat::AbstractArray{ComplexF64} – Matrix used to store the slip in the wavenumber domain

p_rfft::AbstractFFTs.Plan – FFTW plan for forward FFT

p_irfft::AbstractFFTs.Plan – FFTW plan for inverse FFT

Notes

When using GPUs, it is possible to decide where the arrays reside using gpu_id

rfft (https://juliamath.github.io/AbstractFFTs.jl/stable/api/#AbstractFFTs.rfft) is used

to be more memory efficient since we deal only with real numbers (i.e. slip can't be complex).

The FFTW strategy for finding a plan is not defined since cuFFT ignores it so it is useless.

Examples

StressFFT(experiment::AbstractExperiment; gpu_id::Int=0) – Stress law for the current experiment

Error laws

HighSeas.DoubleError — Type

DoubleError <: AbstractErrorLaw (experiment)

Create a DoubleError object used to calculate two errors between matrix A and B:

err (local error): norm(A-B)/norm(A)

erra (absolute error): maximum(abs(A-B))/A[A==maximum(abs(A-B))]

err and erra are returned.

Fields

diff::AbstractArray{Float64} – Matrix to store the difference A-B

absdiff::AbstractArray{Float64} – Matrix to store the abs(A-B)

mask::AbstractArray{Bool} – Matrix to store where A==maximum(abs(A-B))

Notes

When using GPUs, it is possible to decide where the masks reside using gpu_id

Examples

DoubleError(experiment::AbstractExperiment; gpu_id::Int=0) – Create a double error instance for the current experiment