Simulation samplers

Here are the objects that can be used to sample the simulation domain. It is possible to sample either a point (PointSampler), a section (SectionSampler) or contour data (ContourSampler)

HighSeas.PointSampler — Type

PointSampler <: AbstractSampler

Sample simulation at a given point

Fields

mask::AbstractArray{Int8} – Preallocated array used to define the point in the domain (1 only at the coordinate)

NT::Int – Number of samples

dxs::Vector{Float64} – Sampled values of slip (m)

Vs::Vector{Float64} – Sampled values of slip rate (m/s)

thetas::Vector{Float64} – Sampled values of θ

taus::Vector{Float64} – Sampled values of τ (Pa)

times::Vector{Float64} – Sampled times

sample_point_id::Int – Id of the sample point

sample_point_x::Float64 – x coord of the sample point (m)

sample_point_y::Float64 – y coord of the sample point (m)

Notes

When using GPUs, it is possible to decide where the mask reside using gpu_id

It is possible to use a sample point file structured as follows

n x y
1 -36.0 0.0
2 -22.5 -7.50
3 -16.5 -12.0
...

Examples

– PointSampler(sample_point_id::Int, sample_point_x::Float64, sample_point_y::Float64, NT::Int, experiment::AbstractExperiment; gpu_id::Int=0) – Sample at point – PointSampler(sample_points_paths::String, sample_point_id, NT, experiment; gpu_id=0) – Sample point at a poisition given by a txt file

HighSeas.SectionSampler — Type

SectionSampler <: AbstractSampler

Sample simulation using a section

Fields

temp::AbstractArray{Float64} – Temporary matrix used to calculate the section values

mask::AbstractArray{Int8} – Temporary mask used to calculate the section values

NT::Int – Number of section to sample

quantity::Symbol – Which quantity of AbstractState to sample

section::Matrix{Float64} – NT x (n+1) matrix where n is the grid resolution and NT is the number of section to sample

coord::Float64 – Position to sample

axis::String – Which axis

Notes

When using GPUs, it is possible to decide where the mask reside using gpu_id

When using GPUs temp will always be set in unified memory

The first column of the matrix defines the maximum slip (useful to define seismic/aseismic slips)

Examples

– SectionSampler("dx", coord=0.0, axis="y", NT::Int, experiment::AbstractExperiment; gpu_id::Int=0) – Sample slip at all points where y=0.0

HighSeas.ContourSampler — Type

ContourSampler <: AbstractSampler

Sample rupture contours

Fields

contour::AbstractArray{Float64} – Contour matrix

mask::AbstractArray{Int8} – Temporary mask used to calculate the rupture contours

thresh::Float64 – Threshold value for the contour

first_contour::Int8 – Pre allocated matrix used to store the first contour

t_fc::Float64 – First contour time

quantity::Symbol – Quantity to sample

Notes

When using GPUs, it is possible to decide where the mask reside using gpu_id

Examples

– ContourSampler("V", 1e-3, experiment::AbstractExperiment; gpu_id=0) – Sample slip speed when reaching 1e-3